requirements.txt
pytorch-lightning
lightning_bolts
seaborn
argh
jsonargparse[signatures]
flatten_dict
jupyterlab
flake8
yapf
openmim
torch-tb-profiler
pefile